requirements.txt
asgiref==3.8.1
Django==5.1.2
django-ckeditor==6.7.1
django-cors-headers==4.5.0
django-jazzmin==3.0.1
django-js-asset==2.2.0
djangorestframework==3.15.2
drf-yasg==1.21.8
gunicorn==23.0.0
inflection==0.5.1
packaging==24.2
pillow==11.0.0
pytz==2024.2
PyYAML==6.0.2
setuptools==70.1.1
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
